Load past activity when collections list is quite long
ref: https://37s.fizzy.37signals.com/buckets/693169850/bubbles/999008733

This moves the "fetch-on-visible" link to right after the
"div#activity" to which it will append, so that it immediately
triggers if the first day's grid is short.

The link previously was at the end of the page, meaning that a long
list of Collections would push it out of the viewport, requiring a
scroll before loading the next day.